✨ feat(runtime): catalog matcher + 3 resolve diagnostics
filter_tools delegates to AgentSpec::tool_allowed; emits tracing
warns for unmatched, argument misuse, orphan permission rules.
