+++ crates/awaken-runtime/Cargo.toml
@@ -13,6 +13,7 @@ categories = ["artificial-intelligence", "asynchronous"]
 [dependencies]
 awaken-contract = { workspace = true }
 awaken-protocol-a2a = { workspace = true }
+awaken-tool-pattern = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 thiserror = { workspace = true }
